Save initial, working state of xpt
xpt was an ancient X11 application to display a periodic table with
clickable elements.  Years ago I hacked it to show the electronic
structure in addition to the standard data it originally had.

Since then, the C standard has changed and this program no longer
compiled properly -- or rather, it compiled but then segfaulted
because of numerous missing prototypes led to incorrect default use of
integers where non-integers (notably, pointers) were supposed to be.

This commit cleans the program up just barely enough to make it work.
There are still some warnings, but it works, and I want to preserve
the state.

I have done away with the wholly unnecessary Imakefile, and replaced
it with a simple Makefile.  There are very few knobs to twiddle, and
Imake was overkill.  Plus, it didn't work anymore on my system.